changed all html routes to use router
