+++ routes/apiRoutes.js
@@ -1,35 +1,15 @@
+// "use strict";
+
 const db = require("../models");
-const moment = require("moment");
+const express = require("express");
 
-const express = require('express');
 const apiRouter = express.Router();//for api routes
-
-
-const convertTime = body=>{
-    console.log(body);
-    let datetime = moment();
-    if(body.date && body.hour && body.ampm){
-        datetime = moment(body.date+ body.hour + body.ampm, "MMM DD, YYYY hh a");
-    }else if(body.date){
-        datetime = moment(body.date, "MMM DD, YYYY");
-    }
-    return datetime.format("YYYY-MM-DD HH:mm:ss");
-};
-
-const checkAuth = user=>{
-    if(user && user.id && !isNaN(parseInt(user.id))) {//validate user ID
-        return parseInt(user.id);
-        //TODO: check against DB
-    }else{
-        throw 401;//TODO  throw the stack  and proper error
-    }
-};
+const Helper = require("./utils/helper");
 
 // Get all data from current user
 apiRouter.get("/alldata", (req, res, next)=> {
-    console.log("here");
     try{
-        const userId = checkAuth(req.user);
+        const userId = Helper.checkAuth(req.user);
 
         db.User.findOne({
             where:{
@@ -55,7 +35,7 @@ apiRouter.get("/alldata", (req, res, next)=> {
 //post a new stool log for user
 apiRouter.post("/stool", function (req, res, next) {
     try{
-        const userId = checkAuth(req.user);
+        const userId = Helper.checkAuth(req.user);
         if (req.body && req.body.score && !isNaN(parseInt(req.body.score))) {
             db.User.findOne({
                 where: {
@@ -64,7 +44,7 @@ apiRouter.post("/stool", function (req, res, next) {
             }).then(function (userResult) {
                 userResult.createStool({
                     score: req.body.score,
-                    time: convertTime(req.body),
+                    time: Helper.convertTime(req.body),
                     comment: req.body.comment ? req.body.comment : null//TODO:sanitize the input
                 }).then(stool => {
                     res.json(stool);
@@ -81,7 +61,7 @@ apiRouter.post("/stool", function (req, res, next) {
 
 apiRouter.post("/water", function(req, res) {
     try{
-        const userId = checkAuth(req.user);
+        const userId = Helper.checkAuth(req.user);
         if(req.body && req.body.intake && !isNaN(parseInt(req.body.intake))){
             db.User.findOne({
                 where:{
@@ -90,7 +70,7 @@ apiRouter.post("/water", function(req, res) {
             }).then(function(userResult) {
                 userResult.createWater({
                     intake:req.body.intake,
-                    time:convertTime(req.body)
+                    time:Helper.convertTime(req.body)
                 }).then(water=>{
                     res.json(water);
                 });
